requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.3
scikit-learn==1.3.2
numpy==1.26.2
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.2 